+++ src/common/Portal.js
@@ -11,23 +11,6 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {omit} from './utils';
 
-class ChildrenWrapper extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      children: props.children
-    };
-  }
-
-  setChildren(children) {
-    this.setState({children: children});
-  }
-
-  render() {
-    return this.state.children;
-  }
-}
-
 const portalClass = '__portal';
 
 class Portal extends React.Component {
@@ -45,16 +28,7 @@ class Portal extends React.Component {
     document.body.appendChild(portal);
     portal.className = portalClass;
     this.portal = portal;
-    ReactDOM.render(
-      <ChildrenWrapper
-        ref={wrapper => this.wrapper = wrapper}
-      >
-        <div{...this._getChildrenProps(this.props)}>
-          {this.props.children}
-        </div>
-      </ChildrenWrapper>,
-      this.portal
-    );
+    this.renderPortal();
   }
 
   componentWillUnmount() {
@@ -66,18 +40,7 @@ class Portal extends React.Component {
   }
 
   componentDidUpdate() {
-    this.wrapper.setChildren(
-      <div{...this._getChildrenProps(this.props)}>
-        {this.props.children}
-      </div>
-    );
-  }
-
-  _getChildrenProps(props) {
-    return {
-      ...omit(props, ['onDocumentMouseDown', 'onDocumentMouseScroll', 'styles']),
-      style: props.styles
-    };
+    this.renderPortal();
   }
 
   _isDocumentEventOwner(target) {
@@ -96,19 +59,24 @@ class Portal extends React.Component {
     }
   }
 
+  renderPortal() {
+    ReactDOM.render(
+      <div
+        {...omit(this.props, ['onDocumentMouseDown', 'onDocumentMouseScroll'])}
+      >
+        {this.props.children}
+      </div>,
+      this.portal
+    );
+  }
+
   render() {
     return null;
   }
 }
 
-ChildrenWrapper.propTypes = {
-  children: PropTypes.node,
-};
-
 Portal.propTypes = {
   children: PropTypes.node,
-  id: PropTypes.string,
-  styles: PropTypes.object,
   onDocumentMouseDown: PropTypes.func,
   onDocumentMouseScroll: PropTypes.func,
 };